Add parser for timetravel Qurey Language (ttQL)
This new subdirectory and package parsre contains the
self defined TTQL.g4 file which defines the grammer of ttQL.
It is base on OpenCypher. The auto generated files and interfaces
of the the parser and lexer are put in a sub package ttql_listener.
Another sub package "listeners" for listener implementations for the
parser is added along with a first implementation of a listener
for the PropertyLookup clause.