Git refactore parse tree listeners
The PropertyClauseInsight Object for potential Property Lookups is saved
as a map now. The keys are the ComparisonExpressionContext objects, so
all its children are saved to it as a list via its
PropertyClauseInsights.
The ElementListener is introduced. It listenes to Variables of elements
(nodes, edges) occouring in the query. It also checks if it is part
of a MATCh, a WHERE or a RETURN clause.
Some additional information ist listened to. For example, every
comparison counts its PartialComparissonExpressions. Furthermore
it is checked if there is a NullPrdicateExpression. It might be
possible that this only works for one partial comparison expression
yet. A reminder comment is put in place.