Do some cleanup and some restructuring
The existence check of a value for a property lookup comparison
in a WHERE clause does not fetch data anymore but checks the existence
in the timescale database and return just a boolean. This should be
way faster.
Furthermore the calculation of the LookupInfo for the lookup comparions
inside WHERE clause are happening inside the Parser now when iterating
over the insights so it does not have to happen twice. This still could
be done in the listener which would be even better.