docker-compose testing environment
Setup a docker-compose file which with "docker-compose up" starts two testing containers, one
for neo4j and one for timescaledb. Therefore it uses some backup data
which is used for testing purposes. This is to be found in test-data.
The dockerfile uses these files for restoring databases. Now the
integration tests in api_test.go can be run by hand.