Add script for automating the integration test
The script is called by the go test TestIntegration.
It then copies the backup files to the backup volumes
of the docker containers, runs the neo4j backup
(couldnt figure out how to do this in docker compose yet),
runs docker compose, waits until the setup is ready and then
runs the go tests for testing shallow and non-shallow queries.