update and restructure datatabase-api. Unify them among packages
All database access is now done via the databaseapi except for
the only-neo4j data-adpter. this is still to be done. Some renaming
and restructuring of the database-access has been done. Neo4j and
Timescale database access is happening now over dedicated files in
the databaseapi package. A config file config.yaml for the database-access has
been introduced.