Prevent from trying to fetch time-series/properties if they arent
If a property does not start with the prefix properties or ts then
it is not a link to a time-series and nothing can be fetched. So
in this case the timescale querying is skipped, the neo4j property
is left untouched because it contains the complete data in this case.
TODO: differ conditional cases differently on these properties