update data-generator, allow defining data template in a yaml file
The template of the generated data can now be defined in a yaml
file. The data generated will be in the structure described by
the template. Note: the generated data contains a list of labels
now instead of a single label field. We need to adapt this in the
data adapters.