Benchmark
This commit introduces benchmarking on base of different data templates
which generate different datasets. These datasets are then loaded both,
in a only neo4j manner as well as in the ttdb manner to allow performing
equivalent cypher and ttql queries for performance comparison reasons.
The benchmark results are saved to .csv files. (Each query is run 100
times after a warmup of 10 times)