Final commit
I added the raw data of the benchmarks under benchmark/test-data.
furthermore i added the raw results of the response time in a pdf under
benchmark/results. The data of the integration tests are manipulated
within the tests so the tests work again. Since datetimes are stored
as datetime and not as string anymore, this had to be changed for the
integration test data. Right now we do this through Neo4j write
commands. Do a new backup in the future and exchange the backup data.